+++ packages/extension/src/page-side/commit-drivers/select.ts
@@ -1,9 +1,173 @@
-// element-plus select COMMIT page-side driver (IIFE).
-// T2.7a will migrate from dom.ts COMMIT handler L1198-1283.
+// element-plus select COMMIT page-side driver (IIFE, attaches to window.__vortexCommitSelect).
+// Migrated from dom.ts COMMIT handler driverId === "element-plus-select" branch (L1198-1281).
+//
+// Host-side call:
+//   await loadPageSideModule(tabId, frameId, 'commit-select')
+//   const result = await nativePageQuery(tabId, frameId,
+//     (sel, closestSelector, val, timeoutMs) =>
+//       (window as any).__vortexCommitSelect.run(sel, closestSelector, val, timeoutMs),
+//     [selector, driver.closestSelector, value, timeout])
 
 (function () {
-  (window as unknown as { __vortexCommitSelect: Record<string, unknown> }).__vortexCommitSelect = {
-    placeholder: true,
+  if ((window as any).__vortexCommitSelect?.version === 1) return;
+
+  // Page-side inline helpers (matching original dom.ts page-side func; cannot import across files)
+  const sleep = (ms: number) => new Promise((r) => setTimeout(r, ms));
+
+  async function waitFor<T>(
+    probe: () => T | null | undefined,
+    to: number,
+    intervalMs = 50,
+  ): Promise<T | null> {
+    const deadline = Date.now() + to;
+    while (Date.now() < deadline) {
+      const r = probe();
+      if (r) return r;
+      await sleep(intervalMs);
+    }
+    return null;
+  }
+
+  function dispatchMouseClick(el: HTMLElement): void {
+    const rect = el.getBoundingClientRect();
+    const opts = {
+      bubbles: true,
+      cancelable: true,
+      view: window,
+      clientX: rect.left + rect.width / 2,
+      clientY: rect.top + rect.height / 2,
+    };
+    el.dispatchEvent(new MouseEvent("mousedown", opts));
+    el.dispatchEvent(new MouseEvent("mouseup", opts));
+    el.dispatchEvent(new MouseEvent("click", opts));
+  }
+
+  /**
+   * Run el-select commit.
+   * - sel: element selector (the target element passed by agent)
+   * - closestSelector: driver's closestSelector (".el-select")
+   * - val: string or string[] (option label(s) to select)
+   * - timeoutMs: overall operation timeout
+   *
+   * Returns: { result?, error?, errorCode?, stage?, extras? } — same contract as dom.ts COMMIT path.
+   */
+  async function run(
+    sel: string,
+    closestSelector: string,
+    val: unknown,
+    timeoutMs: number,
+  ): Promise<{
+    result?: unknown;
+    error?: string;
+    errorCode?: string;
+    stage?: string;
+    extras?: Record<string, unknown>;
+  }> {
+    // Resolve root element (mirrors dom.ts L939-958 element resolution block)
+    const els = document.querySelectorAll(sel);
+    if (els.length === 0) return { error: `Element not found: ${sel}`, errorCode: "ELEMENT_NOT_FOUND" };
+    if (els.length > 1)
+      return {
+        error: `Selector "${sel}" matched ${els.length} elements`,
+        errorCode: "SELECTOR_AMBIGUOUS",
+        extras: { matchCount: els.length },
+      };
+    const target = els[0] as HTMLElement;
+    const root = (target.closest(closestSelector) ??
+      target.querySelector(closestSelector)) as HTMLElement | null;
+    if (!root)
+      return {
+        error: `Target does not match driver closestSelector "${closestSelector}" (neither ancestor nor descendant)`,
+        errorCode: "UNSUPPORTED_TARGET",
+        extras: { driverId: "element-plus-select" },
+      };
+
+    // -------- Element Plus el-select driver --------
+    const labels = Array.isArray(val)
+      ? (val as unknown[]).map((v) => String(v))
+      : [String(val)];
+    const isMultiple = Array.isArray(val) || root.classList.contains("is-multiple");
+
+    // trigger: el-select 2.x uses .el-select__wrapper, older versions use .select-trigger
+    const wrapper =
+      (root.querySelector(".el-select__wrapper") as HTMLElement | null) ??
+      (root.querySelector(".select-trigger") as HTMLElement | null) ??
+      (root as HTMLElement);
+
+    // 1. Click wrapper to open popper
+    wrapper.scrollIntoView({ block: "center", inline: "center" });
+    dispatchMouseClick(wrapper);
+
+    // 2. Wait for current select's dropdown to appear and be visible.
+    //    el-select wrapper has aria-controls pointing to popper id.
+    const popperId = wrapper.getAttribute("aria-controls");
+    const dropdown = await waitFor(() => {
+      if (popperId) {
+        const el = document.getElementById(popperId);
+        if (el && el.getBoundingClientRect().width > 0) return el as HTMLElement;
+      }
+      // fallback: scan all visible dropdowns, take first
+      const all = document.querySelectorAll(".el-select-dropdown");
+      for (const d of Array.from(all)) {
+        if ((d as HTMLElement).getBoundingClientRect().width > 0) return d as HTMLElement;
+      }
+      return null;
+    }, timeoutMs);
+    if (!dropdown) {
+      return {
+        error: "Select dropdown did not open within timeout",
+        errorCode: "COMMIT_FAILED",
+        stage: "open-dropdown",
+      };
+    }
+
+    // 3. Find and click option for each label
+    const clicked: string[] = [];
+    const unknown: string[] = [];
+    for (const label of labels) {
+      const items = Array.from(
+        dropdown.querySelectorAll(".el-select-dropdown__item"),
+      ) as HTMLElement[];
+      const hit = items.find((it) => (it.textContent || "").trim() === label);
+      if (!hit) {
+        unknown.push(label);
+        continue;
+      }
+      dispatchMouseClick(hit);
+      clicked.push(label);
+      await sleep(40); // let Vue run one tick before clicking next option
+    }
+
+    // 4. Multi-select: click wrapper to close popper; single-select closes automatically
+    if (isMultiple && dropdown.getBoundingClientRect().width > 0) {
+      dispatchMouseClick(wrapper);
+      await sleep(40);
+    }
+
+    if (unknown.length > 0) {
+      const available = Array.from(
+        dropdown.querySelectorAll(".el-select-dropdown__item"),
+      ).map((i) => ((i as HTMLElement).textContent || "").trim());
+      return {
+        error: `Unknown option label(s): ${unknown.join(", ")}. Available: ${available.join(", ")}`,
+        errorCode: "INVALID_PARAMS",
+        extras: { unknown, available },
+      };
+    }
+
+    return {
+      result: {
+        success: true,
+        driver: "element-plus-select",
+        multiple: isMultiple,
+        clicked,
+      },
+    };
+  }
+
+  (window as any).__vortexCommitSelect = {
+    version: 1,
+    run,
   };
 })();
